Swap out WSGI stuff, add __init__.py
The code needed for WSGI is now in an extra file `cards.wsgi` and a
`__init__.py` has been added so that the script can be used as a module.
The extraction of the WSGI code was necessary to be able to use the
Python module without having Flask installed.